move get-printable-host to helpers
